# Reduce allocations when writing