# PR Details

Fix #494

If a row is full, don't bother allocating a new one, just return it.

Use the last populated row as a hint for the size of new rows.

Simplify checkSheet to remove row map (tiny performance improvement, mostly just easier to follow now)

## Related Issue

#494 

## Motivation and Context

Reduce memory consumed while creating a document

## How Has This Been Tested

Added a new benchmark, ensured tests continue to pass

Master
BenchmarkWrite-8   	       2	 601920270 ns/op	207277296 B/op	 1893770 allocs/op

This fix
BenchmarkWrite-8   	       2	 567284257 ns/op	127115028 B/op	 1833744 allocs/op

Rows:
Master
BenchmarkRows-8   	   10000	    362571 ns/op	 2010561 B/op	      30 allocs/op
This fix
BenchmarkRows-8   	   10000	    349607 ns/op	 2010560 B/op	      30 allocs/op

Performance wise, the change I made to rows barely makes a difference, most of that was likely stack allocated anyway. This is just simpler.

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->

- [ ] Docs change / refactoring / dependency upgrade
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist

<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->

- [x] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] I have read the **CONTRIBUTING** document.
- [x] I have added tests to cover my changes.
- [x] All new and existing tests passed.