@mlh758 2019-11-08
I should clarify, the performance of reading a file barely changed which is what I was referring to. This cuts the memory used while writing a file by about 40%.