@mperezguendulain 2020-04-14
I tested the version 1.4.1 and 2.1.0 of excelize, and the writing time improved a lot, the memory use is significantly better, thanks for these, but I think is not enough, because for a lot of information the memory use is a lot too. For example, to write 5 million rows with 16 columns each row the program takes around 4GB of ram.
Here is the example I used: https://gist.github.com/mperezguendulain/393d59f285608fa4b7142ba5120b5109

It will great if it excelize could use a small chunk of memory.